# docs/swagger.yaml
basePath: /api
definitions:
...
  domain.CheckUpdateRequest:
    properties:
      client_info:
        description: 客户端信息，可选
        type: string
      current_version:
        type: string
    required:
    - current_version
    type: object
...
  domain.CreateUpgradeTargetRequest:
    properties:
      description:
        type: string
      name:
        type: string
      package_id:
        type: string
      project_id:
        type: string
      release_id:
        type: string
    required:
    - name
    - package_id
    - project_id
    - release_id
    type: object
...
  domain.Response:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
    type: object
...
  domain.UpdateUpgradeTargetRequest:
    properties:
      description:
        type: string
      is_active:
        type: boolean
      name:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: Package Management System API for managing projects, packages, and
    users with RBAC.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: PKMS API
  version: "1.0"
paths:
...
  /tenant/{id}/user/{userId}/role:
    delete:
      consumes:
      - application/json
      description: Remove a user's role in a specific tenant
      parameters:
      - description: Tenant ID
        in: path
        name: id
        required: true
        type: string
      - description: User ID
        in: path
        name: userId
        required: true
        type: string
      - description: Role information to remove
        in: body
        name: request
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Successfully removed user tenant role
          schema:
            $ref: '#/definitions/domain.Response'
        "400":
          description: Bad request - invalid parameters
          schema:
            $ref: '#/definitions/domain.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/domain.Response'
      security:
      - BearerAuth: []
      summary: Remove user tenant role
      tags:
      - Tenant Users
    get:
      consumes:
      - application/json
      description: Get a user's role in a specific tenant
      parameters:
      - description: Tenant ID
        in: path
        name: id
        required: true
        type: string
      - description: User ID
        in: path
        name: userId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successfully retrieved user tenant role
          schema:
            $ref: '#/definitions/domain.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/domain.Response'
      security:
      - BearerAuth: []
      summary: Get user tenant role
      tags:
      - Tenant Users
  /tenant/{id}/user/role:
    post:
      consumes:
      - application/json
      description: Set a user's role in a specific tenant through Casbin
      parameters:
      - description: Tenant ID
        in: path
        name: id
        required: true
        type: string
      - description: User ID and role information
        in: body
        name: request
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Successfully set user tenant role
          schema:
            $ref: '#/definitions/domain.Response'
        "400":
          description: Bad request - invalid parameters or role
          schema:
            $ref: '#/definitions/domain.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/domain.Response'
      security:
      - BearerAuth: []
      summary: Set user tenant role
      tags:
      - Tenant Users
  /tenant/{id}/users/roles:
    get:
      consumes:
      - application/json
      description: Get all users in a tenant with their role information from Casbin
      parameters:
      - description: Tenant ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successfully retrieved tenant users with roles
          schema:
            $ref: '#/definitions/domain.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/domain.Response'
      security:
      - BearerAuth: []
      summary: Get tenant users with roles
      tags:
      - Tenant Users
...
  /upgrade/check:
    post:
      consumes:
      - application/json
      description: Check if there are available updates for the client
      parameters:
      - description: Update check request
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/domain.CheckUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully checked for updates
          schema:
            $ref: '#/definitions/domain.Response'
        "400":
          description: Bad request - invalid parameters
          schema:
            $ref: '#/definitions/domain.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/domain.Response'
      summary: Check for updates
      tags:
      - Upgrades
  /upgrade/project/{projectId}/targets:
    get:
      consumes:
      - application/json
      description: Get all upgrade targets for a specific project
      parameters:
      - description: Project ID
        in: path
        name: projectId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successfully retrieved project upgrade targets
          schema:
            $ref: '#/definitions/domain.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/domain.Response'
      security:
      - BearerAuth: []
      summary: Get project upgrade targets
      tags:
      - Upgrades
  /upgrade/target:
    post:
      consumes:
      - application/json
      description: Create a new upgrade target
      parameters:
      - description: Upgrade target information
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/domain.CreateUpgradeTargetRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Successfully created upgrade target
          schema:
            $ref: '#/definitions/domain.Response'
        "400":
          description: Bad request - invalid parameters
          schema:
            $ref: '#/definitions/domain.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/domain.Response'
      security:
      - BearerAuth: []
      summary: Create upgrade target
      tags:
      - Upgrades
  /upgrade/target/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a specific upgrade target by ID
      parameters:
      - description: Upgrade target ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successfully deleted upgrade target
          schema:
            $ref: '#/definitions/domain.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/domain.Response'
      security:
      - BearerAuth: []
      summary: Delete upgrade target
      tags:
      - Upgrades
    get:
      consumes:
      - application/json
      description: Get a specific upgrade target by ID
      parameters:
      - description: Upgrade target ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successfully retrieved upgrade target
          schema:
            $ref: '#/definitions/domain.Response'
        "404":
          description: Upgrade target not found
          schema:
            $ref: '#/definitions/domain.Response'
      security:
      - BearerAuth: []
      summary: Get specific upgrade target
      tags:
      - Upgrades
    put:
      consumes:
      - application/json
      description: Update a specific upgrade target by ID
      parameters:
      - description: Upgrade target ID
        in: path
        name: id
        required: true
        type: string
      - description: Upgrade target information
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/domain.UpdateUpgradeTargetRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully updated upgrade target
          schema:
            $ref: '#/definitions/domain.Response'
        "400":
          description: Bad request - invalid parameters
          schema:
            $ref: '#/definitions/domain.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/domain.Response'
      security:
      - BearerAuth: []
      summary: Update upgrade target
      tags:
      - Upgrades
  /upgrade/targets:
    get:
      consumes:
      - application/json
      description: Get all upgrade targets with optional filtering
      parameters:
      - description: Project ID
        in: query
        name: project_id
        type: string
      - description: Package ID
        in: query
        name: package_id
        type: string
      - description: Is active
        in: query
        name: is_active
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: Successfully retrieved upgrade targets
          schema:
            $ref: '#/definitions/domain.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/domain.Response'
      security:
      - BearerAuth: []
      summary: Get upgrade targets
      tags:
      - Upgrades
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"